Set the value of option as uint64_t.

+++ src/Compiler/CompilerOptions.cpp
@@ -46,7 +46,7 @@ bool disableKrnlOpFusion;                              // common for both
 bool disableQuantZeroPoint;                            // common for both
 bool enableKrnlBufferReuse;                            // common for both
 bool disableMemRefPrefetch;                            // common for both
-int compilationNumThreads;                             // common for both
+uint64_t compilationNumThreads;                        // common for both
 EmissionTargetType emissionTarget;                     // onnx-mlir only
 bool invokeOnnxVersionConverter;                       // onnx-mlir only
 bool preserveLocations;                                // onnx-mlir only
@@ -618,7 +618,7 @@ static llvm::cl::opt<bool, true> disableConstantPropOpt("disable-constant-prop",
     llvm::cl::location(disableConstantProp), llvm::cl::init(false),
     llvm::cl::cat(OnnxMlirCommonOptions));
 
-static llvm::cl::opt<int, true> compilation_num_threads("j",
+static llvm::cl::opt<uint64_t, true> compilation_num_threads("j",
     llvm::cl::desc("Use <int> threads for compilation. The default value is "
                    "0, which spawns threads for all available CPUs.\n"),
     llvm::cl::location(compilationNumThreads), llvm::cl::init(0),
